requirements.txt
opencv_python_headless==4.5.3.56
torchsummary==1.5.1
albumentations==1.0.3
pandas==1.3.2
torchvision==0.9.0
numpy==1.20.3
tqdm==4.62.2
matplotlib==3.4.3
umap_learn==0.5.1
seaborn==0.11.2
torch==1.9.0
timm==0.4.12
torchtoolbox==0.1.5
Pillow==8.3.2
scikit_learn==1.0

